+++ .eslintcache
@@ -1,0 +1,1 @@
+[{"/home/sealfury/mid_course_project/flixcloud_client/src/App.jsx":"1","/home/sealfury/mid_course_project/flixcloud_client/src/modules/moviesData.js":"2","/home/sealfury/mid_course_project/flixcloud_client/src/DisplayMoviesData.jsx":"3"},{"size":242,"mtime":1606923569216,"results":"4","hashOfConfig":"5"},{"size":164,"mtime":1606923068050,"results":"6","hashOfConfig":"5"},{"size":970,"mtime":1606923553812,"results":"7","hashOfConfig":"5"},{"filePath":"8","messages":"9","errorCount":0,"warningCount":0,"fixableErrorCount":0,"fixableWarningCount":0},"1sqd6ib",{"filePath":"10","messages":"11","errorCount":0,"warningCount":0,"fixableErrorCount":0,"fixableWarningCount":0},{"filePath":"12","messages":"13","errorCount":0,"warningCount":0,"fixableErrorCount":0,"fixableWarningCount":0},"/home/sealfury/mid_course_project/flixcloud_client/src/App.jsx",[],"/home/sealfury/mid_course_project/flixcloud_client/src/modules/moviesData.js",[],"/home/sealfury/mid_course_project/flixcloud_client/src/DisplayMoviesData.jsx",[]]
